move alias generators to alias_generators.py and add to __all__
