+++ pydantic/decorator.py
@@ -2,11 +2,11 @@ from functools import wraps
 from typing import TYPE_CHECKING, Any, Callable, Dict, List, Mapping, Optional, Tuple, Type, TypeVar, Union, overload
 
 from . import validator
+from .alias_generators import to_camel
 from .config import Extra
 from .errors import ConfigError
 from .main import BaseModel, create_model
 from .typing import get_all_type_hints
-from .utils import to_camel
 
 __all__ = ('validate_arguments',)
